requirements.txt
asgiref==3.8.1
attrs==22.2.0
beautifulsoup4==4.10.0
blinker==1.4
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.0.3
colorama==0.4.4
coverage==7.6.10
crashtest==0.4.1
cryptography==39.0.2
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.1.3
django-adminlte2==0.4.3
django-stdimage==6.0.2
dulwich==0.21.3
evdev==1.4.0
filelock==3.9.0
gunicorn==23.0.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
model-mommy==2.0.0
more-itertools==9.1.0
msgpack==1.0.5
natsort==8.0.2
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pexpect==4.8.0
pillow==11.1.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.19.3
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
setproctitle==1.2.2
shellingham==1.5.0.post1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.2
static3==0.7.0
systemd-python==234
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing_extensions==4.12.2
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
wadllib==1.3.6
webencodings==0.5.1
xdg==5.0.0
zipp==3.15.0
